editors - introduce and use EditorsObserver and use for MRU list
