+++ src/vs/workbench/browser/parts/editor/editor.ts
@@ -151,4 +151,14 @@ export interface EditorServiceImpl extends IEditorService {
 	 * Emitted when an editor failed to open.
 	 */
 	readonly onDidOpenEditorFail: Event<IEditorIdentifier>;
+
+	/**
+	 * Emitted when the list of most recently active editors change.
+	 */
+	readonly onDidMostRecentlyActiveEditorsChange: Event<void>;
+
+	/**
+	 * Access to the list of most recently active editors.
+	 */
+	readonly mostRecentlyActiveEditors: ReadonlyArray<IEditorIdentifier>;
 }
